requirements.txt
numpy
pandas
scikit-image
Pillow
tqdm
torch
torchvision
pyyaml
kornia 
opencv-python 
xgboost
matplotlib